Name the gate response GateDetail and move AnswerGateRequest to schemas
GateView collided with Django's View (the request handler) and was the lone
*View among response DTOs that otherwise read Response/Result/Detail/Status;
GateDetail matches its sibling AgentCallDetail. The request model moves from
the route file to schemas.py, where request bodies live.
